+++ src/main/java/com/example/sistemaBanco/dto/request/PostTransacaoDeposito.java
@@ -1,6 +1,7 @@
 package com.example.sistemaBanco.dto.request;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 import com.example.sistemaBanco.dto.IdDto;
 import com.example.sistemaBanco.entities.Transacao;
@@ -11,12 +12,12 @@ public class PostTransacaoDeposito implements Serializable {
 	private static final long serialVersionUID = 1L;
 
 	private IdDto conta;
-	private Double valor;
+	private BigDecimal valor;
 
 	public PostTransacaoDeposito() {
 	}
 
-	public PostTransacaoDeposito(IdDto conta, Double valor) {
+	public PostTransacaoDeposito(IdDto conta, BigDecimal valor) {
 		super();
 		this.conta = conta;
 		this.valor = valor;
@@ -41,11 +42,11 @@ public class PostTransacaoDeposito implements Serializable {
 		this.conta = conta;
 	}
 
-	public Double getValor() {
+	public BigDecimal getValor() {
 		return valor;
 	}
 
-	public void setValor(Double valor) {
+	public void setValor(BigDecimal valor) {
 		this.valor = valor;
 	}
